# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Medical Agent API
  description: |
    慢病健康管理智能体系统 RESTful API

    提供健康评估、风险预测、个性化处方等功能。
  version: 1.0.0
  contact:
    name: Medical Agent Team
    email: support@medical-agent.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: 开发环境
  - url: https://api.medical-agent.com
    description: 生产环境
...
paths:
...
  /api/v3/skills:
    get:
      tags: [Skills]
      summary: 列出可用 Skills
      operationId: listSkills
      security:
        - ApiKeyAuth: []
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillsListResponse'
  /api/v3/skills/{skill_name}:
    get:
      tags: [Skills]
      summary: 获取 Skill 详情
      operationId: getSkillDetail
      security:
        - ApiKeyAuth: []
      parameters:
        - name: skill_name
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 成功
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SkillDetailResponse'
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: "Format: Bearer YOUR_API_KEY"
  schemas:
...
    SkillsListResponse:
      type: object
      properties:
        skills:
          type: array
          items:
            $ref: '#/components/schemas/Skill'
    Skill:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        category:
          type: string
        description:
          type: string
        enabled:
          type: boolean
        version:
          type: string
        endpoints:
          type: array
          items:
            type: string
    SkillDetailResponse:
      allOf:
        - $ref: '#/components/schemas/Skill'
        - type: object
          properties:
            prompt_template:
              type: string
            model_config:
              type: object
            input_schema:
              type: object
            output_schema:
              type: object
...